+++ Ch14/friends.c
@@ -9,9 +9,9 @@ struct names {
 
 struct guy {
   struct names handle;
-  char favfood[LEN];
-  char job[LEN];
-  float income;
+  char         favfood[LEN];
+  char         job[LEN];
+  double       income;
 };
 
 int main(void) {
